requirements.txt
Appium-Python-Client==3.1.1
asgiref==3.5.2
attrs==23.2.0
blinker==1.4
Brotli @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_38mvgltu8c/croots/recipe/brotli-split_1659616064542/work
brotlipy==0.7.0
certifi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_70oty9s9jh/croot/certifi_1720453497032/work/certifi
cffi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b4nang6w_y/croot/cffi_1700254307954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cryptography==36.0.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==2.0.3
func-timeout==4.3.5
genson==1.2.2
gevent @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6adipc6l_z/croot/gevent_1703058798464/work
greenlet @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_516imz09pb/croot/greenlet_1702059966336/work
h11==0.13.0
h2==4.1.0
hpack==4.0.0
htmlmin @ file:///tmp/build/80754af9/htmlmin_1605825795554/work
hyperframe==6.0.1
idna @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_771olrhiqw/croot/idna_1666125579282/work
importlib-resources==6.1.1
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_369vjbb695/croot/jmespath_1700144575847/work
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kaitaistruct==0.9
ldap3==2.9.1
loguru==0.7.2
MarkupSafe==2.1.3
mitmproxy==8.0.0
msgpack==1.0.8
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
parameterized==0.8.1
passlib==1.7.4
pkgutil_resolve_name==1.3.10
playwright==1.40.0
poium==1.5.1
protobuf==3.19.6
publicsuffix2==2.20191221
pyasn1==0.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyee==11.0.1
PyMySQL==1.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks @ file:///Users/ktietz/Code/oss/ci_pkgs/pysocks_1626781349491/work
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8dd_9u21zz/croot/pyyaml_1670514759576/work
redis @ file:///Users/ktietz/demo/mc3/conda-bld/redis_1629466498261/work
referencing==0.32.1
requests==2.31.0
rpds-py==0.16.2
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
seldom==3.8.0
selenium==4.23.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tornado==6.4.1
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_dc7bd08d-3aa9-4e45-9124-1c432b2bee1dmp2r0ell/croots/recipe/urllib3_1650640006820/work
urwid==2.1.2
webdriver-manager==3.8.5
websocket-client==1.8.0
Werkzeug==2.2.2
wsproto==1.1.0
XTestRunner==1.7.4
zipp==3.17.0
zope.event @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_0c9flhzfuv/croot/zope.event_1702999622126/work
zope.interface @ file:///Users/ktietz/demo/mc3/conda-bld/zope.interface_1628962362010/work
zstandard==0.17.0
